requirements.txt
asgiref==3.6.0
bleach==6.0.0
Django==3.2
django-filter==22.1
django-suit==2.0a2
django-summernote==0.8.20.0
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
